# cmd/portal/docs/swagger.yaml
definitions:
...
  dto.CreateUserReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  dto.GetUserNameResp:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
      users:
        type: string
    type: object
...
  dto.RespError:
    properties:
      error:
        properties:
          code:
            type: integer
          message:
            type: string
        type: object
    type: object
...
  dto.UpdateUserReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /user:
    get:
      description: Get user by username pattern
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: username pattern
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: users info
          schema:
            $ref: '#/definitions/dto.GetUserNameResp'
      summary: Get user by username pattern
      tags:
      - user
    patch:
      description: Update user infomation
      parameters:
      - description: jwt
        in: header
        name: Authorization
        required: true
        type: string
      - description: update Req
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: Update OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Update user infomation
      tags:
      - user
    post:
      description: Create new user
      parameters:
      - description: user info
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: create OK
          schema:
            $ref: '#/definitions/dto.RespError'
      summary: Create new user
      tags:
      - user
swagger: "2.0"